More cleanups

+++ app/src/main/java/io/aether/android/screens/scanner/ScannerScreen.kt
@@ -35,6 +35,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import io.aether.android.R
+import io.aether.android.matter.toInt
 import timber.log.Timber
 
 /**
@@ -121,8 +122,8 @@ fun MatterBeaconItem(beacon: MatterBeacon) {
           text =
               stringResource(
                   R.string.beacon_detail_text,
-                  beacon.vendorId,
-                  beacon.productId,
+                  beacon.vendorId.toInt(),
+                  beacon.productId.toInt(),
                   beacon.discriminator,
               ),
           style = MaterialTheme.typography.bodyMedium,
